bug fixes, new GameState

+++ trunk/tools/Core/AI/Constants.cs
@@ -6,7 +6,7 @@ namespace Core
 {
 	class Constants
 	{
-		public const decimal BombTimeout = 30;
+		public const int BombTimeout = 30;
 		public const int ExplosionDuration = 3;
